Include missing observations as NaNs and not zeros
In Comptabilité Nationale files, blank data is in fact hidden zeros. Parsing
these files amounts in getting either zeros or NaNs for missing data,
depending on if observation exists in the file (but with no numbers) or not.

This commit changes the parsing method in order to consider straight zeros
as NaNs. In this way, if data is missing, formulas using this data will
always be NaN.

Test data is recomputed.

NB : To use the "correct data", you should put option force_recompute = True
when getting the tidy dataframe with the function get_or_construct_data.